split text buffer file in document sub mod
