+++ packages/core/src/document/info.rs
@@ -1,0 +1,135 @@
+use unicode_segmentation::UnicodeSegmentation;
+
+const UTF8_BOM: &str = "\u{feff}";
+
+#[derive(Default, Debug)]
+pub struct TextBufferInfo {
+    encoding: CharacterEncoding,
+    eol: EOL,
+    eos_normalized: bool,
+    // contains_rtl: bool,
+    // contains_unusual_line_terminators: bool,
+    is_ascii: bool,
+    // normalize_eol: bool,
+}
+
+impl TextBufferInfo {
+    pub fn new(default_eol: DefaultEOL) -> TextBufferInfo {
+        let eol = match default_eol {
+            DefaultEOL::LF => EOL::LF,
+            DefaultEOL::CRLF => EOL::CRLF,
+        };
+
+        TextBufferInfo {
+            encoding: CharacterEncoding::Utf8,
+            eol,
+            ..Default::default()
+        }
+    }
+
+    pub fn from_string(value: &String, default_eol: DefaultEOL) -> (TextBufferInfo, Vec<i32>) {
+        let encoding = CharacterEncoding::from(value);
+        let mut line_starts = vec![0];
+        let mut line_break_count = LineBreakCount::default();
+        let mut is_ascii = true;
+
+        let enumerate = &mut value.grapheme_indices(true).enumerate();
+        while let Some((i, (grapheme_index, c))) = enumerate.next() {
+            match c {
+                "\r" => match enumerate.nth(i + 1) {
+                    Some((_, (grapheme_index, "\n"))) => {
+                        line_starts.push(grapheme_index as i32);
+                        line_break_count.crlf += 1;
+                    }
+                    Some(_) => {
+                        line_starts.push(grapheme_index as i32);
+                        line_break_count.cr += 1;
+                    }
+                    None => {
+                        line_starts.push(grapheme_index as i32);
+                        line_break_count.cr += 1;
+                    }
+                },
+                "\n" => {
+                    line_starts.push(grapheme_index as i32);
+                    line_break_count.lf += 1;
+                }
+                _ => {
+                    if !c.is_ascii() {
+                        is_ascii = false
+                    }
+                }
+            }
+        }
+
+        let total_eol_count = line_break_count.cr + line_break_count.lf + line_break_count.crlf;
+        let total_cr_count = line_break_count.cr + line_break_count.crlf;
+
+        let eol = match (total_eol_count, default_eol) {
+            (x, default_eol) if x == 0 && default_eol == DefaultEOL::LF => EOL::LF,
+            (x, _) if x == 0 || total_cr_count > total_eol_count / 2 => EOL::CRLF,
+            _ => EOL::LF,
+        };
+
+        let info = TextBufferInfo {
+            encoding,
+            eol,
+            eos_normalized: false,
+            is_ascii,
+        };
+
+        (info, line_starts)
+    }
+
+    pub fn should_check_crlf(&self) -> bool {
+        return !(self.eos_normalized && self.eol == EOL::LF);
+    }
+}
+
+#[derive(PartialEq)]
+pub enum DefaultEOL {
+    LF = 1,
+    CRLF = 2,
+}
+
+#[derive(Debug, PartialEq)]
+enum EOL {
+    LF,
+    CR,
+    CRLF,
+}
+
+impl Default for EOL {
+    fn default() -> EOL {
+        EOL::LF
+    }
+}
+
+#[derive(Debug)]
+enum CharacterEncoding {
+    Utf8,
+    Utf8WithBom,
+}
+
+impl Default for CharacterEncoding {
+    fn default() -> CharacterEncoding {
+        CharacterEncoding::Utf8
+    }
+}
+
+impl From<&String> for CharacterEncoding {
+    fn from(s: &String) -> Self {
+        if s.starts_with(UTF8_BOM) {
+            CharacterEncoding::Utf8WithBom
+        } else {
+            CharacterEncoding::Utf8
+        }
+    }
+}
+
+#[derive(Debug, Default)]
+struct LineBreakCount {
+    cr: i32,
+    lf: i32,
+    crlf: i32,
+}
