+++ packages/core/src/document/text_buffer.rs
@@ -1,33 +1,12 @@
-use bevy::ecs::prelude::*;
-use intrusive_collections::intrusive_adapter;
-use intrusive_collections::{KeyAdapter, RBTree, RBTreeAtomicLink};
-use std::{convert::From, fs, str};
+use crate::document::{
+    buffer::{Buffer, BufferCursor},
+    cache::TextBufferCache,
+    info::{DefaultEOL, TextBufferInfo},
+    node::{Node, NodePosition},
+    piece::{NodeAdapter, Piece},
+};
+use intrusive_collections::RBTree;
 use unicode_segmentation::UnicodeSegmentation;
-
-#[derive(Component, Debug)]
-pub struct Document {
-    file_path: Option<&'static str>,
-    text_buffer: TextBuffer,
-}
-
-impl Document {
-    pub fn new(default_eol: DefaultEOL) -> Document {
-        Document {
-            file_path: None,
-            text_buffer: TextBuffer::new(default_eol),
-        }
-    }
-
-    pub fn from_file(file_path: &'static str, default_eol: DefaultEOL) -> Document {
-        let original = fs::read_to_string(file_path.clone()).expect("Failed to read file");
-        let text_buffer = TextBuffer::from_string(&original, default_eol);
-
-        Document {
-            file_path: Some(file_path),
-            text_buffer,
-        }
-    }
-}
 
 #[derive(Default, Debug)]
 pub struct TextBuffer {
@@ -41,14 +20,14 @@ pub struct TextBuffer {
 }
 
 impl TextBuffer {
-    fn new(default_eol: DefaultEOL) -> TextBuffer {
+    pub fn new(default_eol: DefaultEOL) -> TextBuffer {
         TextBuffer {
             info: TextBufferInfo::new(default_eol),
             ..Default::default()
         }
     }
 
-    fn from_string(original: &str, default_eol: DefaultEOL) -> TextBuffer {
+    pub fn from_string(original: &str, default_eol: DefaultEOL) -> TextBuffer {
         let original = original.to_string();
         let (info, line_starts) = TextBufferInfo::from_string(&original.clone(), default_eol);
 
@@ -63,9 +42,7 @@ impl TextBuffer {
 
         text_buffer
     }
-}
 
-impl TextBuffer {
     pub fn insert(&mut self, offset: i32, value: &'static str) {
         let value = value.to_string();
 
@@ -393,23 +370,58 @@ impl TextBuffer {
     }
 
     fn recompute_tree_metadata(&mut self, node: &Node) {
-        let mut cursor = self.tree.find(&node.total_len());
-        let current = cursor.get().expect("Node provided is null");
-        if self.is_root(&current) {
-            return;
-        }
+        // let delta = 0;
+        // let lf_delta = 0;
+        // if (node === self.tree.root) {
+        //     return;
+        // }
 
-        while let Some(node) = cursor.peek_parent().peek_right().get() {
-            if current.total_len() == node.total_len() {
-                cursor.move_parent();
-            }
-        }
+        // // go upwards till the node whose left subtree is changed.
+        // while (x !== tree.root && x === x.parent.right) {
+        //     x = x.parent;
+        // }
 
-        if self.is_root(current) {
-            return;
-        }
+        // if (x === tree.root) {
+        //     // well, it means we add a node to the end (inorder)
+        //     return;
+        // }
 
-        cursor.move_parent();
+        // // x is the node whose right subtree is changed.
+        // x = x.parent;
+
+        // delta = calculateSize(x.left) - x.size_left;
+        // lf_delta = calculateLF(x.left) - x.lf_left;
+        // x.size_left += delta;
+        // x.lf_left += lf_delta;
+
+        // // go upwards till root. O(logN)
+        // while (x !== tree.root && (delta !== 0 || lf_delta !== 0)) {
+        //     if (x.parent.left === x) {
+        //         x.parent.size_left += delta;
+        //         x.parent.lf_left += lf_delta;
+        //     }
+
+        //     x = x.parent;
+        // }
+
+        // original
+        // let mut cursor = self.tree.find(&node.total_len());
+        // let current = cursor.get().expect("Node provided is null");
+        // if self.is_root(&current) {
+        //     return;
+        // }
+
+        // while let Some(node) = cursor.peek_parent().peek_right().get() {
+        //     if current.total_len() == node.total_len() {
+        //         cursor.move_parent();
+        //     }
+        // }
+
+        // if self.is_root(current) {
+        //     return;
+        // }
+
+        // cursor.move_parent();
 
         // let left = cursor.peek_left().get();
         // if let Some(node) = cursor.get() {
@@ -516,484 +528,12 @@ impl TextBuffer {
     }
 }
 
-#[derive(Debug, Clone)]
-struct Buffer {
-    value: String,
-    line_starts: Vec<i32>,
-}
-
-impl Default for Buffer {
-    fn default() -> Buffer {
-        Buffer {
-            value: Default::default(),
-            line_starts: vec![0],
-        }
-    }
-}
-
-impl From<String> for Buffer {
-    fn from(value: String) -> Buffer {
-        let line_starts = Buffer::get_line_starts(&value.clone());
-        Buffer { value, line_starts }
-    }
-}
-
-impl Buffer {
-    fn new(value: String, line_starts: Vec<i32>) -> Buffer {
-        Buffer { value, line_starts }
-    }
-
-    fn offset(&self, cursor: BufferCursor) -> i32 {
-        self.line_starts.get(cursor.line as usize).unwrap_or(&0) + cursor.column
-    }
-
-    fn get_line_starts(value: &String) -> Vec<i32> {
-        let mut line_starts = vec![0];
-
-        let enumerate = &mut value.grapheme_indices(true).enumerate();
-        while let Some((i, (grapheme_index, c))) = enumerate.next() {
-            match c {
-                "\r" => {
-                    enumerate.nth(i + 1);
-                    line_starts.push(grapheme_index as i32 + 1);
-                }
-                "\n" => {
-                    line_starts.push(grapheme_index as i32 + 1);
-                }
-                _ => {}
-            }
-        }
-
-        line_starts
-    }
-
-    fn get_line_feed_count(&self, start: &BufferCursor, end: &BufferCursor) -> i32 {
-        if end.column == 0 || end.line == self.line_starts.len() as i32 - 1 {
-            return end.line - start.line;
-        }
-
-        let end_line = end.line as usize;
-        let next_line_start_offset = self.line_starts[end_line + 1];
-        let end_offset = self.line_starts[end_line] + end.column;
-        if next_line_start_offset > end_offset + 1 {
-            return end.line - start.line;
-        }
-
-        let previous_grapheme_offset = end_offset as usize - 1;
-        if self.value.graphemes(true).collect::<Vec<&str>>()[previous_grapheme_offset] == "\r" {
-            return end.line - start.line + 1;
-        } else {
-            return end.line - start.line;
-        }
-    }
-}
-
-const UTF8_BOM: &str = "\u{feff}";
 const CHANGE_NODE_DOES_NOT_EXIST: &str = "Change node doesn't exist";
-
-#[derive(Default, Debug)]
-pub struct TextBufferInfo {
-    encoding: CharacterEncoding,
-    eol: EOL,
-    eos_normalized: bool,
-    // contains_rtl: bool,
-    // contains_unusual_line_terminators: bool,
-    is_ascii: bool,
-    // normalize_eol: bool,
-}
-
-impl TextBufferInfo {
-    fn new(default_eol: DefaultEOL) -> TextBufferInfo {
-        let eol = match default_eol {
-            DefaultEOL::LF => EOL::LF,
-            DefaultEOL::CRLF => EOL::CRLF,
-        };
-
-        TextBufferInfo {
-            encoding: CharacterEncoding::Utf8,
-            eol,
-            ..Default::default()
-        }
-    }
-
-    fn from_string(value: &String, default_eol: DefaultEOL) -> (TextBufferInfo, Vec<i32>) {
-        let encoding = CharacterEncoding::from(value);
-        let mut line_starts = vec![0];
-        let mut line_break_count = LineBreakCount::default();
-        let mut is_ascii = true;
-
-        let enumerate = &mut value.grapheme_indices(true).enumerate();
-        while let Some((i, (grapheme_index, c))) = enumerate.next() {
-            match c {
-                "\r" => match enumerate.nth(i + 1) {
-                    Some((_, (grapheme_index, "\n"))) => {
-                        line_starts.push(grapheme_index as i32);
-                        line_break_count.crlf += 1;
-                    }
-                    Some(_) => {
-                        line_starts.push(grapheme_index as i32);
-                        line_break_count.cr += 1;
-                    }
-                    None => {
-                        line_starts.push(grapheme_index as i32);
-                        line_break_count.cr += 1;
-                    }
-                },
-                "\n" => {
-                    line_starts.push(grapheme_index as i32);
-                    line_break_count.lf += 1;
-                }
-                _ => {
-                    if !c.is_ascii() {
-                        is_ascii = false
-                    }
-                }
-            }
-        }
-
-        let total_eol_count = line_break_count.cr + line_break_count.lf + line_break_count.crlf;
-        let total_cr_count = line_break_count.cr + line_break_count.crlf;
-
-        let eol = match (total_eol_count, default_eol) {
-            (x, default_eol) if x == 0 && default_eol == DefaultEOL::LF => EOL::LF,
-            (x, _) if x == 0 || total_cr_count > total_eol_count / 2 => EOL::CRLF,
-            _ => EOL::LF,
-        };
-
-        let info = TextBufferInfo {
-            encoding,
-            eol,
-            eos_normalized: false,
-            is_ascii,
-        };
-
-        (info, line_starts)
-    }
-
-    fn should_check_crlf(&self) -> bool {
-        return !(self.eos_normalized && self.eol == EOL::LF);
-    }
-}
-
-#[derive(PartialEq)]
-pub enum DefaultEOL {
-    LF = 1,
-    CRLF = 2,
-}
-
-#[derive(Debug, PartialEq)]
-enum EOL {
-    LF,
-    CR,
-    CRLF,
-}
-
-impl Default for EOL {
-    fn default() -> EOL {
-        EOL::LF
-    }
-}
-
-#[derive(Debug)]
-enum CharacterEncoding {
-    Utf8,
-    Utf8WithBom,
-}
-
-impl Default for CharacterEncoding {
-    fn default() -> CharacterEncoding {
-        CharacterEncoding::Utf8
-    }
-}
-
-impl From<&String> for CharacterEncoding {
-    fn from(s: &String) -> Self {
-        if s.starts_with(UTF8_BOM) {
-            CharacterEncoding::Utf8WithBom
-        } else {
-            CharacterEncoding::Utf8
-        }
-    }
-}
-
-#[derive(Debug, Default)]
-struct LineBreakCount {
-    cr: i32,
-    lf: i32,
-    crlf: i32,
-}
-
-#[derive(Debug)]
-struct PieceTreeSearchCache {
-    limit: i32,
-    positions: Vec<NodePosition>,
-    line_positions: Vec<NodeLinePosition>,
-}
-
-impl Default for PieceTreeSearchCache {
-    fn default() -> PieceTreeSearchCache {
-        PieceTreeSearchCache {
-            limit: 1,
-            positions: vec![],
-            line_positions: vec![],
-        }
-    }
-}
-
-impl PieceTreeSearchCache {
-    fn get_position(&self, offset: i32) -> Option<&NodePosition> {
-        let mut res = None;
-        for p in self.positions.iter().rev() {
-            if p.node_start_offset <= offset && p.node_start_offset + p.node.piece.len >= offset {
-                res = Some(p);
-                break;
-            }
-        }
-
-        res
-    }
-
-    // fn get_line_position(&self, line_number: i32) -> Option<&NodeLinePosition> {
-    //     let mut res = None;
-    //     for p in self.line_positions.iter().rev() {
-    //         if p.node_start_line_number < line_number
-    //             && p.node_start_line_number + p.node.piece.line_feed_count >= line_number
-    //         {
-    //             res = Some(p);
-    //             break;
-    //         }
-    //     }
-
-    //     res
-    // }
-
-    fn is_limit(&self) -> bool {
-        self.positions.len() + self.line_positions.len() >= self.limit as usize
-    }
-
-    fn set_position(&mut self, position: NodePosition) {
-        if self.is_limit() {
-            self.positions.remove(0);
-        }
-        self.positions.push(position);
-    }
-
-    // fn set_line_position(&mut self, line_position: NodeLinePosition) {
-    //     if self.is_limit() {
-    //         self.line_positions.remove(0);
-    //     }
-    //     self.line_positions.push(line_position);
-    // }
-}
-
-#[derive(Default, Debug)]
-struct TextBufferCache {
-    last_change: BufferCursor,
-    last_visited: LineCache,
-    search: PieceTreeSearchCache,
-    line_count: i32,
-    len: i32,
-}
-
-#[derive(Default, Debug)]
-struct LineCache {
-    line_number: i32,
-    value: String,
-}
-
-impl LineCache {
-    fn new(line_number: i32, value: String) -> LineCache {
-        LineCache { line_number, value }
-    }
-}
-
-#[derive(Debug, Clone)]
-struct NodePosition {
-    node: Node,
-    remainder: i32,
-    node_start_offset: i32,
-}
-
-impl NodePosition {
-    fn new(node: Node, remainder: i32, node_start_offset: i32) -> NodePosition {
-        NodePosition {
-            node,
-            remainder,
-            node_start_offset,
-        }
-    }
-}
-
-#[derive(Debug)]
-struct NodeLinePosition {
-    node: Node,
-    node_start_offset: i32,
-    node_start_line_number: i32,
-}
-
-#[derive(Default, Debug, Clone)]
-pub struct Node {
-    link: RBTreeAtomicLink,
-
-    piece: Piece,
-    r#type: NodeType,
-
-    left_len: i32,
-    left_line_feed_count: i32,
-}
-
-impl Node {
-    fn is_original(&self) -> bool {
-        self.r#type == NodeType::Original
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-enum NodeType {
-    Original,
-    Changed,
-}
-
-impl Default for NodeType {
-    fn default() -> NodeType {
-        NodeType::Changed
-    }
-}
-
-impl Node {
-    fn from_original_buffer(buffer: &Buffer) -> Node {
-        let grapheme_len = buffer.value.graphemes(true).count() as i32;
-        let line_starts_len = buffer.line_starts.len() as i32;
-        let end_line = if line_starts_len == 0 {
-            0
-        } else {
-            line_starts_len - 1
-        };
-
-        Node {
-            piece: Piece::new(
-                BufferCursor::default(),
-                BufferCursor::new(end_line, grapheme_len),
-                line_starts_len,
-                grapheme_len,
-            ),
-            r#type: NodeType::Original,
-            ..Default::default()
-        }
-    }
-
-    fn from_changed_buffer(buffer: &Buffer, start_offset: i32, start: BufferCursor) -> Node {
-        let end_offset = buffer.value.graphemes(true).count() as i32;
-        let line_starts_len = buffer.line_starts.len() as i32;
-        let end_line = if line_starts_len == 0 {
-            0
-        } else {
-            line_starts_len - 1
-        };
-        let end_column = end_offset - buffer.line_starts[end_line as usize];
-        let end = BufferCursor::new(end_line as i32, end_column);
-
-        Node {
-            piece: Piece::new(
-                start,
-                end,
-                buffer.get_line_feed_count(&start, &end),
-                end_offset - start_offset,
-            ),
-            r#type: NodeType::Changed,
-            ..Default::default()
-        }
-    }
-
-    fn total_len(&self) -> i32 {
-        self.left_len + self.piece.len
-    }
-
-    fn start_with_lf_from_string(value: &String) -> bool {
-        return value.as_str().graphemes(true).last() == Some("\n");
-    }
-
-    fn end_with_cr_from_string(value: &String) -> bool {
-        match value.graphemes(true).last() {
-            Some(c) => match c {
-                "\r" => true,
-                _ => false,
-            },
-            None => false,
-        }
-    }
-
-    fn start_with_lf(&self, buffer: &Buffer) -> bool {
-        if buffer.line_starts.len() == 0 {
-            return false;
-        } else {
-            if self.piece.start.line == buffer.line_starts.len() as i32 - 1 {
-                return false;
-            }
-
-            let next_line_offset = buffer.line_starts[self.piece.start.line as usize + 1];
-            let start_offset =
-                buffer.line_starts[self.piece.start.line as usize] + self.piece.start.column;
-            if next_line_offset > start_offset + 1 {
-                return false;
-            }
-
-            return buffer.value.graphemes(true).nth(start_offset as usize) == Some("\n");
-        }
-    }
-}
-
-#[derive(Default, Debug, Clone)]
-pub struct Piece {
-    start: BufferCursor,
-    end: BufferCursor,
-    line_feed_count: i32,
-    len: i32,
-}
-
-impl Piece {
-    fn new(start: BufferCursor, end: BufferCursor, line_feed_count: i32, len: i32) -> Piece {
-        Piece {
-            start,
-            end,
-            line_feed_count,
-            len,
-        }
-    }
-}
-
-intrusive_adapter!(pub NodeAdapter = Box<Node>: Node { link: RBTreeAtomicLink });
-impl<'a> KeyAdapter<'a> for NodeAdapter {
-    type Key = i32;
-    fn get_key(&self, node: &'a Node) -> i32 {
-        node.total_len()
-    }
-}
-
-impl From<Piece> for Node {
-    fn from(piece: Piece) -> Self {
-        Self {
-            piece,
-            ..Default::default()
-        }
-    }
-}
-
-#[derive(Default, Debug, Clone, Copy)]
-pub struct BufferCursor {
-    line: i32,
-    column: i32,
-}
-
-impl BufferCursor {
-    fn new(line: i32, column: i32) -> Self {
-        Self { line, column }
-    }
-}
 
 #[cfg(test)]
 mod inserts_and_deletes {
-    use crate::buffer::{DefaultEOL, TextBuffer};
+    use crate::document::{info::DefaultEOL, text_buffer::TextBuffer};
+
     #[test]
     fn basic_insert_and_delete() {
         let mut buffer =
